* rm read method from frame
