+++ baseio-codec/src/main/java/com/generallycloud/baseio/codec/http11/WebSocketFrame.java
@@ -15,28 +15,17 @@
  */
 package com.generallycloud.baseio.codec.http11;
 
-import java.io.IOException;
-
-import com.generallycloud.baseio.buffer.ByteBuf;
+import com.generallycloud.baseio.collection.FixedThreadStack;
 import com.generallycloud.baseio.common.Encoding;
+import com.generallycloud.baseio.component.NioEventLoop;
 import com.generallycloud.baseio.component.NioSocketChannel;
 import com.generallycloud.baseio.protocol.AbstractFrame;
 import com.generallycloud.baseio.protocol.Frame;
 
 public class WebSocketFrame extends AbstractFrame implements HttpMessage {
 
-    public static final int    OP_CONTINUATION_FRAME     = 0;
-    public static final int    OP_TEXT_FRAME             = 1;
-    public static final int    OP_BINARY_FRAME           = 2;
-    public static final int    OP_CONNECTION_CLOSE_FRAME = 8;
-    public static final int    OP_PING_FRAME             = 9;
-    public static final int    OP_PONG_FRAME             = 10;
-    public static final int    HEADER_LENGTH             = 2;
-    public static final String CHANNEL_KEY_SERVICE_NAME  = "CHANNEL_KEY_SERVICE_NAME";
-
     private byte[]             byteArray;
     private boolean            eof;
-    private int                limit;
     private String             readText;
     private String             serviceName;
     private byte               type;
@@ -44,12 +33,7 @@ public class WebSocketFrame extends AbstractFrame implements HttpMessage {
     public WebSocketFrame() {
         this.type = WebSocketCodec.TYPE_TEXT;
     }
-
-    public WebSocketFrame(NioSocketChannel ch,int limit) {
-        this.limit = limit;
-        this.setServiceName(ch);
-    }
-
+    
     public byte[] getByteArray() {
         return byteArray;
     }
@@ -75,104 +59,46 @@ public class WebSocketFrame extends AbstractFrame implements HttpMessage {
     }
 
     public boolean isCloseFrame() {
-        return OP_CONNECTION_CLOSE_FRAME == type;
+        return type == WebSocketCodec.OP_CONNECTION_CLOSE_FRAME;
     }
     
     public boolean isContinuationFrame(){
-        return OP_CONTINUATION_FRAME == type;
+        return type == WebSocketCodec.OP_CONTINUATION_FRAME;
     }
 
     public boolean isEof() {
         return eof;
     }
 
-    @Override
-    public boolean read(NioSocketChannel ch, ByteBuf src) throws IOException {
-        if (src.remaining() < 2) {
-            return false;
+    @SuppressWarnings("unchecked")
+    public void release(NioEventLoop eventLoop) {
+        //FIXME ..final statck is null or not null
+        if (WebSocketCodec.WS_PROTOCOL_CODEC.getFrameStackSize() == 0) {
+            return;
         }
-        src.markP();
-        byte b0 = src.getByte();
-        byte b1 = src.getByte();
-        int dataLen = 0;
-        boolean hasMask = (b1 & 0b10000000) > 0;
-        if (hasMask) {
-            dataLen += 4;
+        FixedThreadStack<WebSocketFrame> stack = (FixedThreadStack<WebSocketFrame>) eventLoop
+                .getAttribute(WebSocketCodec.FRAME_STACK_KEY);
+        if (stack != null) {
+            stack.push(this);
         }
-        int payloadLen = (b1 & 0x7f);
-        if (payloadLen < 126) {
+    }
 
-        } else if (payloadLen == 126) {
-            dataLen += 2;
-            if (src.remaining() < dataLen) {
-                src.resetP();
-                return false;
-            }
-            payloadLen = src.getUnsignedShort();
-        } else {
-            dataLen += 8;
-            if (src.remaining() < dataLen) {
-                src.resetP();
-                return false;
-            }
-            payloadLen = (int) src.getLong();
-            if (payloadLen < 0) {
-                throw new IOException("over limit:" + payloadLen);
-            }
-        }
-        if (payloadLen > limit) {
-            throw new IOException("over limit:" + payloadLen);
-        }
-        if (src.remaining() < payloadLen) {
-            src.resetP();
-            return false;
-        }
-        eof = (b0 & 0b10000000) > 0;
-        byte type = (byte) (b0 & 0xF);
-        if (type == WebSocketCodec.TYPE_PING) {
-            setPing();
-        } else if (type == WebSocketCodec.TYPE_PONG) {
-            setPong();
-        }else{
-            this.type = type;
-        }
-        byte[] array = new byte[payloadLen];
-        if (hasMask) {
-            byte m0 = src.getByte();
-            byte m1 = src.getByte();
-            byte m2 = src.getByte();
-            byte m3 = src.getByte();
-            src.get(array);
-            int length = array.length;
-            int len = (length / 4) * 4;
-            for (int i = 0; i < len;) {
-                array[i++] ^= m0;
-                array[i++] ^= m1;
-                array[i++] ^= m2;
-                array[i++] ^= m3;
-            }
-            if (len < length) {
-                int i = len;
-                int remain = length - len;
-                if (remain == 1) {
-                    array[i++] ^= m0;
-                }else if(remain == 2) {
-                    array[i++] ^= m0;
-                    array[i++] ^= m1;
-                }else {
-                    array[i++] ^= m0;
-                    array[i++] ^= m1;
-                    array[i++] ^= m2;
-                }
-            }
-        } else {
-            src.get(array);
-        }
-        this.byteArray = array;
-        if (type == WebSocketCodec.TYPE_BINARY) {
-            // FIXME 处理binary
-        }
-        return true;
+    protected WebSocketFrame reset(String serviceName, int limit) {
+        this.byteArray = null;
+        this.eof = false;
+        this.readText = null;
+        this.type = 0;
+        this.serviceName = serviceName;
+        super.reset();
+        return this;
+    }
+
+    public void setByteArray(byte[] byteArray) {
+        this.byteArray = byteArray;
+    }
+
+    public void setEof(boolean eof) {
+        this.eof = eof;
     }
 
     @Override
@@ -187,8 +113,16 @@ public class WebSocketFrame extends AbstractFrame implements HttpMessage {
         return super.setPong();
     }
 
+    public void setReadText(String readText) {
+        this.readText = readText;
+    }
+    
     protected void setServiceName(NioSocketChannel ch) {
-        this.serviceName = (String) ch.getAttribute(CHANNEL_KEY_SERVICE_NAME);
+        this.serviceName = (String) ch.getAttribute(WebSocketCodec.CHANNEL_KEY_SERVICE_NAME);
+    }
+
+    protected void setServiceName(String serviceName) {
+        this.serviceName = serviceName;
     }
 
     protected void setWsType(byte type) {
@@ -198,31 +132,6 @@ public class WebSocketFrame extends AbstractFrame implements HttpMessage {
     @Override
     public String toString() {
         return getReadText();
-    }
-
-    //    @Override
-    //    public void release(NioEventLoop eventLoop) {
-    //        super.release(eventLoop);
-    //        //FIXME ..final statck is null or not null
-    //        if (WebSocketCodec.WS_PROTOCOL_CODEC.getFrameStackSize() == 0) {
-    //            return;
-    //        }
-    //        FixedThreadStack<WebSocketFrameImpl> stack = (FixedThreadStack<WebSocketFrameImpl>) eventLoop
-    //                .getAttribute(WebSocketCodec.FRAME_STACK_KEY);
-    //        if (stack != null) {
-    //            stack.push(this);
-    //        }
-    //    }
-
-    protected WebSocketFrame reset(NioSocketChannel ch, int limit) {
-        this.byteArray = null;
-        this.eof = false;
-        this.readText = null;
-        this.type = 0;
-        this.limit = limit;
-        this.setServiceName(ch);
-        super.reset();
-        return this;
     }
 
 }
