+++ baseio-test/src/main/java/com/generallycloud/test/io/load/http11/TestHttpLoadServer.java
@@ -16,10 +16,10 @@
 package com.generallycloud.test.io.load.http11;
 
 import com.generallycloud.baseio.buffer.ByteBuf;
+import com.generallycloud.baseio.codec.http11.HttpCodec;
+import com.generallycloud.baseio.codec.http11.HttpFrame;
 import com.generallycloud.baseio.codec.http11.HttpHeader;
 import com.generallycloud.baseio.codec.http11.HttpStatic;
-import com.generallycloud.baseio.codec.http11.ServerHttpCodec;
-import com.generallycloud.baseio.codec.http11.ServerHttpFrame;
 import com.generallycloud.baseio.component.ChannelAcceptor;
 import com.generallycloud.baseio.component.ChannelContext;
 import com.generallycloud.baseio.component.IoEventHandle;
@@ -36,7 +36,7 @@ public class TestHttpLoadServer {
 
             @Override
             public void accept(NioSocketChannel channel, Frame frame) throws Exception {
-                ServerHttpFrame f = (ServerHttpFrame) frame;
+                HttpFrame f = (HttpFrame) frame;
                 f.setResponseHeader(HttpHeader.Content_Type_Bytes, HttpStatic.plain_bytes);
                 f.setResponseHeader(HttpHeader.Server_Bytes, null);
                 frame.write("Hello World", channel);
@@ -50,9 +50,10 @@ public class TestHttpLoadServer {
         NioEventLoopGroup group = new NioEventLoopGroup();
         group.setMemoryPoolCapacity(1024 * 64);
         group.setMemoryPoolUnit(512);
+        group.setEventLoopSize(4);
         ChannelContext context = new ChannelContext(8087);
         ChannelAcceptor acceptor = new ChannelAcceptor(context, group);
-        context.setProtocolCodec(new ServerHttpCodec(4));
+        context.setProtocolCodec(new HttpCodec(1));
         context.setIoEventHandle(eventHandle);
         context.addChannelEventListener(new LoggerChannelOpenListener());
